Combined script for generating Python classes from abi
